Adding null checks and move JobManagerConstants to api
