+++ titus-server-master/src/main/java/com/netflix/titus/master/jobmanager/service/event/JobModelReconcilerEvent.java
@@ -22,7 +22,7 @@ import com.netflix.titus.api.jobmanager.model.CallMetadata;
 import com.netflix.titus.api.jobmanager.model.job.Job;
 import com.netflix.titus.common.framework.reconciler.EntityHolder;
 import com.netflix.titus.common.framework.reconciler.ModelActionHolder;
-import com.netflix.titus.master.jobmanager.service.JobManagerConstants;
+import com.netflix.titus.api.jobmanager.service.JobManagerConstants;
 import com.netflix.titus.master.jobmanager.service.common.action.TitusChangeAction;
 
 public abstract class JobModelReconcilerEvent extends JobManagerReconcilerEvent {
@@ -33,7 +33,6 @@ public abstract class JobModelReconcilerEvent extends JobManagerReconcilerEvent 
 
     public static class JobNewModelReconcilerEvent extends JobModelReconcilerEvent {
         private final EntityHolder newRoot;
-
         public JobNewModelReconcilerEvent(EntityHolder newRoot) {
             super(newRoot.getEntity(), "-1", (CallMetadata)newRoot.getAttributes().get(JobManagerConstants.JOB_MANAGER_ATTRIBUTE_CALLMETADATA));
             this.newRoot = newRoot;
